feat: share chat state across pipelines
